requirements.txt
apache-airflow
dvc
beautifulsoup4
requests
dvc-gdrive